requirements.txt
python-dotenv==1.0.0
pytest==7.4.2
pytest-benchmark==4.0.0
pytest-html==4.0.2
Faker==19.6.2
tqdm=4.66.1
pandas==2.1.1
snowflake-connector-python==3.2.0